Added A Card Template to Project Component
Created a card template in the Project component. Cards will show a picture of the product and have a small overlay to provide quick information to the user when they hover over the card. The cards are clickable, opening a link to the project in a new tab.

Also added an image of my first project (Shardrun) to use for the card.